Added Semilattice and BoundedSemilattice.
I also tried to provide some initial documentation for lattices,
but I'm sure the text can be improved. There is some tension between
using mathematical symbols (concise and easy to read IMO) and using
the methods (denser but closer to code).

It would be still be good to link to the relevant Wikipedia articles and
other useful sources as well.