Add serializability test for lattices.
This commit adds tests to ensure lattices (and semilattices)
are all serializable. It also adds a min/max lattice for
integers to use in the lattice test.

One downside to mixing all the algebraic structures that
apply to a type into a "default algebra" is that it means
our serializability test can pass even if the traits in
question lack Serializable.