Remove Boolean, Byte, and Short specialization from most things.
The only places where we still specialize on these types is in
Eq/PartialOrder/Order, where we specialize on everything.
This is because these type classes have much wider usage, and
also because it's likely that users will want to define instances
for case classes and other compound types which might want to
make use of specialized instances for Char, Byte, etc.