Implement basic statistical test for hashing function.
This commit introduces a HashTest that can be used to ensure
that a hashing function's output is uniformly distributed
with 95% confidence.

It also introduces an instance of Hash[Long], which uses
Knuth's MMIX RNG algorithm, as a test case.

One proviso is that as-written, Order[A] and Hash[A] conflict
(they both provide `on` methods). Since Order and Hash both
extend Eq, it's not possible to provide separate instances
without conflicting implicits.

For now, I have commented-out Hash#on to avoid these issues.
We should discuss what the correct solution is before merging
this PR.