This commit adds a ton of instances for the type classes.
It also introduces a util.StaticMethods object for more
involved implementations that can be shared across instances.

We had talked about getting rid of some of the more "numbery"
type clases but this PR supports all of them. I can amend it
if necessary but it seemed like a bad idea to combine this
with that pruning effort.