Add Unit instances
Mostly I wanted the `Order` instance for `Unit`. The use-case is that in
some tests we want to check if something two instances of
`Error Xor Unit` are equal, and the derived `Eq` instance for `Xor`
needs an `Eq` instance for both sides.

I have no idea whether the `CommutativeRig[Unit]` is a good idea or is
useful. It seems to be law-abiding though. Let me know if you want me to
rip it out.