Simplify laws, remove platform-specific macros, etc.
Since we started algebra, the catalysts project (which is
a test dependency of ours) has adopted both the Platform
object and the TypeTagM class. We no longer need to define
these in Algebra to use them in our tests.

This commit removes one of the three remaining subprojects
(macros), and simplifies the source structure of laws so
it no longer needs shared/jvm/js subdirectories.