Add Rüdiger and Tomas to the list of maintainers.
I nominated both of them, the existing maintainers were unanimous in
agreeing to offer maintainership to them, and they both accepted.

I've also increased number of sign-offs needed (N) from 1 to 2.

With seven maintainers we should be able to meet this requirement,
even if some of us are less active than others (for many significant
PRs we've already done this).

If this ends up being too much of a burden we can reduce N back to 1.